+++ Vidly/Vidly/Models/MemberShipType.cs
@@ -7,10 +7,11 @@ namespace Vidly.Models
 {
     public class MembershipType
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
         [StringLength(1000)]
         public string Name { get; set; }
+
         public short SignUpFee { get; set; }
         public byte DurationInMonths { get; set; }
         public byte DiscountRate { get; set; }
